New podman::manage_container type for container unit
Assuming at least podman 4.4.0 the new type `podman::mange_container`
will create a quadlet unit file a container which will inturn
generate a service unit for managing the container.

Currently only a tiny subset of a container unit is possible and
only a root container can be run.

The included examples are operational however.

This can be extended later to support a fuller container description
as well rootless containers and also units for pods, networks and
volumes.